requirements.txt
# Core numerical stack
numpy>=1.26
pandas>=2.1
scipy>=1.11

# Market data
yfinance>=0.2

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Config management
pyyaml>=6.0

# File formats
pyarrow>=14.0

# Testing
pytest>=7.4

# Optional (if you add EWMA / stats helpers)
statsmodels>=0.14

# Optional (Monte Carlo reproducibility utilities)
tqdm>=4.66